SEPM project/core/semantic_matcher.py
import logging
from sentence_transformers import SentenceTransformer
from sklearn.metrics.pairwise import cosine_similarity
import numpy as np

# Load BERT model for semantic similarity
# all-MiniLM-L6-v2 is an extremely fast and highly effective model for semantic text matching
logging.info("Loading SentenceTransformer model (all-MiniLM-L6-v2)...")
try:
    model = SentenceTransformer('all-MiniLM-L6-v2')
except Exception as e:
    logging.error(f"Failed to load sentence-transformers model: {e}")
    model = None

def _calculate_keyword_overlap(text1, text2):
    """
    Simple keyword overlap fallback when AI models fail to load.
    Returns a score between 0 and 1.
    """
    words1 = set(text1.lower().split())
    words2 = set(text2.lower().split())
    
    if not words1 or not words2:
        return 0.0
        
    intersection = words1.intersection(words2)
    # Jaccard similarity or simple overlap
    return len(intersection) / min(len(words1), len(words2))

def calculate_match_score(resume_text, job_description):
    """
    Calculates the deep semantic similarity between a resume and a job description.
    Returns a float score between 0 and 100.
    """
    if not resume_text or not job_description:
        return 0.0
        
    # FALLBACK: If ML model didn't load (SSL/Network error), use keyword matching
    if not model:
        overlap_score = _calculate_keyword_overlap(resume_text, job_description)
        # Boost keywords slightly so they pass thresholds
        return round(min(100.0, overlap_score * 150), 2)
        
    try:
        resume_embedding = get_embedding(resume_text).reshape(1, -1)
        job_embedding = get_embedding(job_description).reshape(1, -1)
        
        similarity = cosine_similarity(resume_embedding, job_embedding)[0][0]
        
        # Convert from [-1, 1] to [0, 100] scale
        # Text cosine similarity typically ranges from [0, 1] for embeddings
        score = max(0.0, min(100.0, similarity * 100))
        return round(score, 2)
    except Exception:
        # Final safety fallback
        return round(_calculate_keyword_overlap(resume_text, job_description) * 100, 2)
