SEPM project/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.1
spacy==3.7.2
sentence-transformers>=2.4.0
pdfplumber==0.10.3
pandas==2.1.3
numpy==1.26.2
google-generativeai==0.8.3
